Prevent OOM when setting region cube
When setting region cube, make sure player has not set two positions very far apart.

This can happen when player has been editing one part of the map, then travels to a distant part of the map and tries to start setting a new region.

This does not affect the ability to operate on large areas, nor the setting of the markers, just the region cube marking.