Remove unused variables and assignments
The removal of unused variables can improve the detection of mistakes with luacheck;
for example, it can detect if the wrong argument is passed to a function because of a copy-paste mistake.

Setting `src_data` and `dst_data` to `nil` in `worldedit.stack2`,
which might enable an earlier garbage collection,
looks like an unneeded micro-optimisation.